backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
pandas==2.2.3
numpy==2.1.2
psycopg2-binary==2.9.9 ; platform_system != "Windows"
python-dateutil==2.9.0
